+++ routes/review.js
@@ -6,6 +6,7 @@ const wrapAsync = require("../utils/wrapAsync.js");
 const ExpressError = require("../utils/ExpressError.js");
 // For server side validation (joi)
 const {reviewSchema} = require("../schema.js");
+const { isLoggedIn } = require("../middleware.js");
 
 
 // For review validation
@@ -21,11 +22,12 @@ const validateReview = (req, res, next) => {
 
 
 // Post Review route
-router.post("/", validateReview, wrapAsync(async (req, res) => {
+router.post("/", isLoggedIn, validateReview, wrapAsync(async (req, res) => {
     let { id } = req.params;
     let listing = await Listing.findById(id);
     // let { rating: newRating, comment: newComment } = req.body.review;
     let newReview = new Review(req.body.review)
+    newReview.author = req.user._id;
 
     listing.reviews.push(newReview);
 
@@ -41,9 +43,14 @@ router.post("/", validateReview, wrapAsync(async (req, res) => {
 
 
 // Delete Review Route
-router.delete("/:reviewId", wrapAsync(async (req, res) => {
+router.delete("/:reviewId", isLoggedIn, wrapAsync(async (req, res) => {
     let { id, reviewId } = req.params;
 
+    let isReviewAuthor = await Review.findById(reviewId);
+    if (!isReviewAuthor.author.equals(res.locals.currUser._id)) {
+        req.flash("error", "You are not author of this review.");
+        return res.redirect(`/listings/${id}`);
+    }
     await Listing.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
     await Review.findByIdAndDelete(reviewId);
